+++ heuristics/heuristics_intervals.py
@@ -2,7 +2,7 @@ import numpy as np
 from datetime import datetime
 from os import path, makedirs
 
-from imp_marl.struct_env import Struct
+from imp_env.struct_env import Struct
 
 
 class Heuristics():
