+++ imp_env/gym_sa_struct.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from gym import spaces
 
-from imp_marl.struct_env import Struct
+from imp_env.struct_env import Struct
 
 
 class GymSaStruct(gym.Env):
