+++ wrappers/test_wrappers/check_both_env_owf.py
@@ -1,6 +1,6 @@
 import numpy as np
-from imp_marl.pymarl_ma_struct import PymarlMAStruct
-from imp_marl.pymarl_sa_struct import PymarlSAStruct
+from imp_env.pymarl_ma_struct import PymarlMAStruct
+from imp_env.pymarl_sa_struct import PymarlSAStruct
 import os
 
 if __name__ == '__main__':
